re added a few methods to support usage in script repository
